requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

keyboard==0.13.5
matplotlib==3.9.2
numpy==2.1.2
pillow==11.0.0
PyAudio==0.2.14
pystray==0.19.5
SpeechRecognition==3.11.0
